+++ src/components/Navbar.jsx
@@ -1,11 +1,15 @@
-import { useNavigate, Link } from "react-router-dom";
+import { useNavigate, Link, useLocation } from "react-router-dom";
 import { signOut } from "firebase/auth";
 import { auth } from "../apis/firebase";
 import styled from "styled-components";
-import { useState } from "react";
+import { useEffect, useState } from "react";
 import { LoginPage } from "../pages";
+import axios from "axios";
+import QueryString from "qs";
 
 const Navbar = () => {
+  const location = useLocation();
+  const history = useNavigate();
   const REST_API_KEY = "06264d97cddc6d0d5ef77a0f28d69af9";
   const REDIRECT_URI = "http://localhost:3000/";
   const link = `https://kauth.kakao.com/oauth/authorize?client_id=${REST_API_KEY}&redirect_uri=${REDIRECT_URI}&response_type=code`;
@@ -17,6 +21,61 @@ const Navbar = () => {
   console.log(userNickName);
 
   const [showModal, setShowModal] = useState(false);
+
+  //kakaologin get location
+  const KAKAO_CODE = location.search.split("=")[1];
+  console.log(KAKAO_CODE);
+  //   getuser 실행
+
+  const REST_API_KEY_KAKAO = "06264d97cddc6d0d5ef77a0f28d69af9";
+  const REDIRECT_URI_KAKAO = "http://localhost:3000/";
+  const link_kakao = `https://kauth.kakao.com/oauth/authorize?client_id=${REST_API_KEY_KAKAO}&redirect_uri=${REDIRECT_URI_KAKAO}&response_type=code`;
+  const CLIENT_SECRET = "jvRkvzZgcAhb2iq42YyYwqCoIY5t1uXS";
+  const [nickName, setNickName] = useState();
+  const [profileImage, setProfileImage] = useState();
+  const [accessToken, setAccessToken] = useState();
+
+  console.log(accessToken);
+  const getUser = async () => {
+    const ACCESS_TOKEN = await fetch("https://kauth.kakao.com/oauth/token", {
+      method: "POST",
+      headers: {
+        "Content-Type": "application/x-www-form-urlencoded;charset=utf-8",
+      },
+      body: QueryString.stringify({
+        grant_type: "authorization_code",
+        client_id: REST_API_KEY_KAKAO,
+        redirect_uri: REDIRECT_URI_KAKAO,
+        code: KAKAO_CODE,
+        client_secret: CLIENT_SECRET,
+      }),
+    })
+      .then((res) => res.json())
+      .catch((error) => console.log("error:", error));
+
+    console.log("ACCESS_TOKEN1", ACCESS_TOKEN);
+    setAccessToken(ACCESS_TOKEN.access_token);
+    console.log("ACCESS_TOKEN2", ACCESS_TOKEN.access_token);
+    localStorage.setItem("token_for_kakaotalk", ACCESS_TOKEN.access_token);
+
+    const user = await axios.get("https://kapi.kakao.com/v2/user/me", {
+      headers: {
+        Authorization: `Bearer ${ACCESS_TOKEN.access_token}`,
+      },
+    });
+
+    console.log(user);
+    setNickName(user.data.properties.nickname);
+    setProfileImage(user.data.properties.profile_image);
+    sessionStorage.setItem("id", user.data.properties.nickname);
+  };
+  console.log(nickName, profileImage);
+
+  useEffect(() => {
+    getUser();
+  }, []);
+  //  userdata를 세션이나 로컬 스토리지에 담아준다
+  // 또는 유즈이펙트로 감지하여 실행해준다
 
   // 로그아웃
   const LogOutHandler = async () => {
@@ -31,23 +90,25 @@ const Navbar = () => {
       method: "POST",
     }).then((res) => res.json());
 
-    console
-      .log("isLogout", islogout)
-      .then(() => {
-        alert("로그아웃 되었습니다.");
+    console.log("isLogout", islogout);
+    //   .then(() => {
+    //     alert("로그아웃 되었습니다.");
 
-        // 로그아웃 성공
-        setShowModal(false);
-        navigate("/", { replace: true });
-      })
-      .catch((error) => {
-        // 로그아웃 실패
-        alert("로그아웃에 실패했습니다.");
-      });
+    //     // 로그아웃 성공
+    //     setShowModal(false);
+    //     navigate("/", { replace: true });
+    //   })
+    //   .catch((error) => {
+    //     // 로그아웃 실패
+    //     alert("로그아웃에 실패했습니다.");
+    //   });
     sessionStorage.removeItem("id");
-    // window.location.reload();
+    localStorage.removeItem("token_for_kakaotalk");
+    navigate("/");
+    window.location.reload();
   };
-
+  const localId = sessionStorage.getItem("id");
+  console.log(localId);
   return (
     <Nav>
       <LeftSection>
@@ -61,9 +122,9 @@ const Navbar = () => {
       </MenuSection>
       <InfoSection>
         <LoginButtonBox>
-          {currentUsers !== null ? (
+          {localId !== null ? (
             <>
-              <div>{currentUsers}님</div>
+              <div>{localId}님</div>
               <LoginButton onClick={LogOutHandler}>Logout</LoginButton>
             </>
           ) : (
